Initial Commit
Initial commit of project. The project only contains a main.cpp and
a very simple console GUI in ConsoleGUI class to allow abstracted
simple console I/O.

The code in main is only test code at this stage and there is no
actual functionality to the code at this point.