Add DataLoader Class
DataLoader Class added to the aplication to allow .csv bathy data
files to be loaded into the program. The file data will be returned
from the LoadDataFile() as a 2D std::vector. One point is that the
loader currently expects the bathy data to be rectangular with even
sample spacing and ordered by UTM eastings and then UTM northing. It
also expects the .csv to be in the standard UK Hydrograpic Office
format (UTM_Eastings, UTM_Northings, Depth).

Data files sorted as per DataLoader requirments and plots added as
reference. DataNotes presentation also added to keep track of note.