Add CsvWriter Class
The CsvWriter Class has been added to allow easy writting of .csv
files. Currently this class only has the one function WriteToCsv()
which allows the writing of single UTM Coords to a file or a
std::vector of UTM Coords. This should allow the easy export of
start and end point and the path from each algorithm.