Implement A* Pathfinding Functionality
A* Pathfinding added to find the path between start and finish nodes
the path is then exported as a .csv file of UTM coordinates of the
path. The path has been implemented to only move in 4-directions as
the data is on a grid for now. The heuristic selected to estimate the
cost to the goal is the Manhattan Distance. The A* algorithm is
configured to look for the shortest, smoothest path. This is achieved
by taking into account the distance cost of each movement but also
including the difference in depth between each node.

To allow for the implementation of the A* algorithm a custom implementation
of a priority queue has been added in the PriorityQueue Class. This inserts
the items in order of F_Cost. The implementation maintains the FIFO ordering
of nodes where nodes are equal.