Add Weighting To Lee Pathfinding
The LeeNodes now include a WeightedDistance that take into account
the difference in depth between this node and the previous. When run
this seems to produce the same resuls as generic Lee algorithm. This
does not seem to be an implementation issue. However, more of a look
into this is needed.