Post Submission Close Out
This is a tidy up commit to store a final version of the tool with
all the information, scripts and results stored in one place.

This commit also contains a final version of the slides used to
present the results of this algorithm comparison along with details
of the data set and tool's functionality.